requirements-worker.txt
celery>=5.6.3
redis>=7.4.0
supabase>=2.28.3
python-dotenv>=1.2.2
boto3>=1.42.78
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.28
unstructured[all-docs]
